+++ appinio-search.php
@@ -22,15 +22,6 @@ if (! defined('ABSPATH')) {
     exit;
 }
 
-if (! defined('APPINIO_CDN_URL')) {
-    \define('APPINIO_CDN_URL', 'https://cdn.app-in.io/v1/search.js');
-}
-
-\define('APPINIO_PLUGIN_FILE', __FILE__);
-\define('APPINIO_PLUGIN_DIR', plugin_dir_path(__FILE__));
-\define('APPINIO_PLUGIN_URL', plugin_dir_url(__FILE__));
-\define('APPINIO_VERSION', '0.9.0');
-
 require_once __DIR__ . '/autoload.php';
 
 add_action('before_woocommerce_init', function (): void {
@@ -54,5 +45,5 @@ add_action('plugins_loaded', function (): void {
         return;
     }
 
-    AppInIo\Plugin::instance()->boot();
+    AppInIo\Plugin::instance()->boot(__FILE__);
 });